started doing the to_be_sent ppt

+++ grade10/quarter1/20_09/5-6/6.py
@@ -1,4 +1,4 @@
-x, y = map(float, input().split())
+x, y = map(int, input().split())
 
 if (y <= x ** 2 and x >= 2 - x) or (y <= x ** 2 and y <= 4 - x ** 2 and y <= 2 - x and x <= 0) or (y <= 2 - x and y <= x ** 2 and y >= 0) or (y <= x ** 2 and y >= 4 - x ** 2 and x >= 1):
     print('YES')
